Assuming that no one has tried to upgrade from 1.0 yet (reasonable assumption, as I even endorsed it being broken), let's add this column without doing a new update_N() function.
The column is called "type" (one more of those) and the solution to the
one problem where I couldn't imagine how the search query would look like.
With this "type" flag, we can have both the real "member" items of
an operation as well as "fake" cached source items that should also be
searched for. For get_operation_items(), we're going to skip the cached
items and only get the real ones.

Also, remove the TODOs and get_affected_item(), as those are now
definitely obsolete and fulfill no purpose in HEAD anymore.