The CVS backend's centralize-item-revision-tables update function was still not good enough as it didn't cover the empty item revision object that is required for joining "added" actions (naturally without source items) to the {item_revisions} table nevertheless.
This commit adds such an "empty" item revision to the .install file
so that versioncontrol_add_source_items() works correctly.

And because we want to use 0 as type for that new empty item, take the
chance to sanitize the constants and let them all start at 1. No usage
to be ported in the API module itself, backends need to increase the
ITEM_*, FILE_* and ACTION_* constants by 1 if they stored them in the
database. (Will be taken care of for CVS and SVN in a moment.)