New method for retrieving file contents: now people should just get a copy. The main gripe with the previous version was of course that the whole file was required to be loaded into memory, which is just sooo evil.
I had first tried to make streams, but I dropped that again because of
1. caller comfort (easier to transfer and copy around), and
2. we need files anyways for reliable mimetype detection which is
   crucial for anything that involves file transfers and telling apart
   text and binary files (something that each repository viewer must be
   able to do).

Specifying the 'binary' attribute in the function itself is only supported
by a small subset of version control systems (I don't believe this is
supported by any VCS other than SVN) and even in that case very unreliable
to retrieve the mimetype.

So, whole file copies. Enjoy, and witness the related repoview commit!